error vehiculo corregido, falta ver Depositos

+++ src/main/java/com/example/logistic/model/roles/Vehiculo.java
@@ -23,11 +23,11 @@ public class Vehiculo {
 
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
